Update solvers.md to use comparable solvers
NewtonRhapson is not a good choice because there is a near singularity in the Jacobian, so the different methods are relying on using a Gauss-Newton-like representation to handle that effectively. But NonlinearSolve is for some reason hardcoded to Newton, which is pretty much never a good choice for this package, so this shows something that wouldn't be suggested to users. When this is corrected then the example is fine.